requirements.txt
google-search-results
python-dotenv
requests
pandas
openpyxl
